Pull basic LMD Ghost into fork choice crate.
- Moves the basic implementation into the fork choice crate.
- Builds the option of fork choices into beacon_struct.
