+++ beacon_node/beacon_chain/Cargo.toml
@@ -13,6 +13,7 @@ failure = "0.1"
 failure_derive = "0.1"
 genesis = { path = "../../eth2/genesis" }
 hashing = { path = "../../eth2/utils/hashing" }
+fork_choice = { path = "../../eth2/fork_choice" }
 parking_lot = "0.7"
 log = "0.4"
 env_logger = "0.6"
