+++ beacon_node/beacon_chain/src/lib.rs
@@ -3,7 +3,6 @@ mod attestation_targets;
 mod beacon_chain;
 mod block_graph;
 mod checkpoint;
-mod lmd_ghost;
 
 pub use self::beacon_chain::{BeaconChain, Error};
 pub use self::checkpoint::CheckPoint;
