+++ eth2/fork_choice/Cargo.toml
@@ -8,4 +8,3 @@ edition = "2018"
 db = { path = "../../beacon_node/db" }
 ssz = { path = "../utils/ssz" }
 types = { path = "../types" }
-
